started adding unit analysis (in progress)

+++ core/src/org/neubauerfelix/manawars/manawars/entities/IActionUser.kt
@@ -6,7 +6,7 @@ import org.neubauerfelix.manawars.game.entities.IEntity
 interface IActionUser: ILooking, ITeamable, IEntity, IAnimated {
 
 
-    val actionCooldown: Long
+    val actionCooldown: Float
     fun executeAction(): Boolean
     fun canPerformAction(): Boolean
